requirements.txt
pandas
numpy
seaborn
scikit-learn
tensorflow-macos
tensorflow-metal
gradio==4.14.0
huggingface_hub==0.23.4
-e .